Switched to Cloudflare.  Finished database access guide.
There had been a problem with bootstrap tabs and timedatepicker, and
when the files are served by the Cloudflare CDN they work fine.  I
suspect this is something in the response entities, but was unable to
correct it.  This way does have the advantage of lowering the network
load on the AQA site.

Added more images and instructions to database guide.  Converted it to
'pill' tabs for navigating between sections.  Added instructions for
making CSV files from database queries.