Fixed security problem.
Before: A user could modify the configuration of a machine in another
institution, and then click "Customize Plan".  The changes to the
machine would be saved, although the user would be not be allowed to
create a custom plan.

After: If a user clicks "Customize Plan" for a machine from another institution, then no changes to the machine will be saved and they will not be allowed to create a custom plan.