working on problems with access to static content
The problem was that when static content was viewed, the authentication
challenge was reset - not sure how.

I 'think' that the problem was that the staticDirRestlet was being
created in a lazy way, and when the instance was referenced (in def
getRequestedRole) that it had not yet been created and so messed things
up.  I added a specific reference to it (which caused it to be
constructed) after the server was started but before a client could
reference it.  This way when getRequestedRole was invoked, it would
actually be there.  This seems to work and I got the AWS site up with it
but I intend to look at it more closely.