Solid fix for aliases not showing up properly.
Problem was that AnonymousTranslate needed to be secure, not publik.  If
it was set to public (in WebServer) then it would mess up the
credentials and no one would be able to get the anonymous translate
table so they could not see real IDs.

Also required fix in AQA.js so that if user was asking for static
content then it would not try to get the anonymous translate table. This
is a thing when a publik user comes in and just wants to look at the
user guides (which is completely legit) but keeps getting annoying and
confusing and unnecessary prompts for login.