do alias translation multiple times
Sometimes with a large page, the page would not be completely loaded
when the js had run.  This resulted in the latter part of the page not
getting the alias values updated.  The fix was to do the alias
translation at page load, then more times at intervals to catch any HTML
that was only available later.