Fixed baseline problem.  Added filtering for output and machines.
Baseline problem was that if a Redo on an output was done, the old
baseline no longer had an associated output, which it was using for the
acquisition (data) date.  Fix was to use the acquisitionDate in the
Baseline record instead, and not rely on the output record at all.

There is some question of definition here: If a Redo is performed,
should the corresponding baseline values be updated?  It is conundrums
like this that makes me (Jim) question baselines.

Also added filtering to the Machine and Output list so that users would
only see entries from their own institutions.  Need to add GUI
components to give them more flexibity to see others' data.