Changed algorithm for finding leaf sides.
Now instead of using the best 2 sides closest to the side being
adjusted, it picks the best side in the lower third of the leaves and
the best in the upper third and always uses those.  What sometimes
happened was that the best leaves were clustered all near the top or
bottom of the range, so any error was amplified for leaves that were far
away.