Fixed EPID charts to not be 'flattened' by an out of bounds value.
Occasionally a bad value is included in the data.  When this happens it
can be way out of bounds, and the chart will expand to show it.  When
this happens, the in-bounds values are scaled down so far that they look
like a flat line.  These changed limit the height of the chart in an
intelligent way.