enforce Await=true flag
When using this flag, now only one process will be allowed to run at a
time.  If a client request comes in with Await=true while there is
already another process running with Await=true, then the second process
will wait for the first to complete.

The Await=true also will not return to the caller until processing has
been completed (for that specific request).

This is supported for BBbyCBCT, BBbyEPID, and Phase2 when performing a
'run'.  It is not supported for 'Redo'.  It is also not supported for
LOC.