Limit the number of values shown in chart.
Controlled by configuration value CenterDoseHistoryRange.

Hyejoo Kang from LUHS had an extreme change, and the graph came out as a
single 'step' function.  Changed the limit on the AWS platform to 18
which works for LUHS, but there should be a better solution, such as
allowing a sliding window view across the chart. Reference email 13 July
2020.