Finished fixing DicomSeries so that everything is in the database.
Put all RTPLANs from the share directory into the database.
Technically, the shared directory is no longer needed, but (except for
clutter) it does not hurt to keep it.

Some of the machines could not be determined from the DICOM, so machine
was left null for those.

Used admin user for each institution because it could not be determined
from the DICOM.

The fix utilities in DicomSeries could probably be removed, but (except
for clutter) it does not hurt to keep it.