Added columns to Daily QA CSV
Embedded DICOM metadata into BBbyCBCT and BBbyEPID so that it would be
quick to access for CSV reporting.

Storing the data that way is a compromise.  Accessing it is quick and
easy programatically, but pretty much impossible via SQL.

This commit has the code that fixes the use of the
BBbyCBCT.cbctSeriesInstanceUid field which was being filled with the
SOPInstanceUID.  The code also populates the embedded DICOM for both
BBbyCBCT and BBbyEPID, which will be done as part of regular processing
in the future.