Added extra check to CBCT BB coarse location.
Problem was that noise would sometimes have the right number of 'lit'
pixels to look like the cube when looking for the top of the cube.
Solution was to add a check that looked for the largest rises and falls
of the pixel profiles in the X and Z direction, and require that the
approximate distance between them was the same as the size of the cube.