Reorganized writePng to operate faster in a multi-core environment.
Separated the rendering from the file writing.  The writing of of the
file only takes about 6% of the time of the rendering.  The rendering
does not need to be thread safe, but the file writing does.

Of the course the 6% figure is dependent on the disk speed and image
size.