More error tolerance.
When looking for matching edge pairs, allow them to be interupted by
another edge pair.

When grabbing the slice, allow it to not contain the cube.

Base band proximity directly on a fixed number of voxels.

When checking whether a slice is the top most one, require there to be
several consecutive non-cube slices above it.