Added factor to allow for possibility of bright noise in CBCT Daily QA.
Sometimes there are bright voxels that are not part of the cube, either
due to something else on the table or noise.  If the brightest pixels
are removed to leave only the cube, then it is possible instead to
remove part of the cube and keep the bright non-cube voxels.