Added caching of PatientProcedureXml to reduce load on server.
The XML text is now cached.  It is invalidated when a change is made to the PatientProcedure list.  Since the contents do not change often, this reduces the server load because it can usually return the cached value as opposed to looking things up in the database.

Also changed PatientProcedureUpdate so that it invalidated the anonymous cache.