When there are multiple instances of the same beam, then use the one that was captured last.
(e.g.: Two "T2 Open" beams) use the one with the latest Acquisition date+time.  The assumption
is that the beam was re-captured because of some problem with the earlier beam(s).  This also
gives operators control so that they can fix a mistake.  If the first one were used, then they
could not fix the mistake.

Making the assumption automatically also lets processing proceed with no user input.