Added fix for cleaning up redundant DICOM series.
Occasionally there is more than one instance of the same DICOM series stored, usually because there was a crash of some sort.

The fix was to add code that is run at startup that will remove them.  They are removed by removing the inputs that they reference.  Because of a DB constraint set to CASCADE, the DICOM series is deleted as well.