Deprecated use of InstanceCreationDate and InstanceCreationTime.
They often indicate the time that the file was C-MOVE'ed, not when it was created at the treatment machine.

In particular, when a DICOM C-MOVE is done with the Varian VMSDBD daemon, it sets these values to the current date and time.  To be explicit: Performing the same C-MOVE again will yield a new date and time.