Add check for Winston Lutz box being too small.
Previously WL would just report an unexpected error to the user.  Now it checks the box size, and, if it is too small, rejects it and tells the user what the problem is.

If the box is too small, then the penumbras of box edges and ball interfere with each other, and it diminishes the accuracy of the measurement.