IntelliJ IDEA configuration change to use immutable.Seq
New version of Scala compiler does not assume that Seq is immutable.

The code was written with the assumption that Seq is immutable.

This configuration change makes it explicit that the immutable version is to be used.

This image was captured to allow a developer in the future to see what is going on and how to modify it if needed.  This is especially important because it is not otherwise obvious that it is needed or how it is set up.