Keep cached copy of metadata to improve consistency and performance.
When using the cache, get it once and use that same copy for the entire duration of that use.  This ensures consistency (otherwise metadata might change during that use).

Cache the metadata for the entire service (Singleton pattern, one location) so that it does not have to be regenerated for each use.  This improves performance.