Check for the existence of an Output row prior to insert.
Sometimes a referenced Output row would not exist prior to inserting a data row (for BBbyEPID and OutputFile), even though it had been created.  This resulted in a constraint error.

These fixes are intended to test for and wait for the output to be created prior to inserting the data that references it.  (This should not be required.)  It is difficult to test this problem because it happens very infrequently.  I hope it works.

This problem has only been observed in SqlServer, not PostgreSQL.

From the log on UMRO production:

Error from Slick database connection: com.microsoft.sqlserver.jdbc.SQLServerException: The MERGE statement conflicted with the FOREIGN KEY constraint "BBbyEPID_outputPKConstraint". The conflict occurred in database "AQAmsPR", table "dbo.output", column 'outputPK'.

Error from Slick database connection: com.microsoft.sqlserver.jdbc.SQLServerException: The MERGE statement conflicted with the FOREIGN KEY constraint "BBbyEPID_outputPKConstraint". The conflict occurred in database "AQAmsPR", table "dbo.output", column 'outputPK'.